+++ source/test-suite/fixtures/InstanceHeadExpected.hs
@@ -4,3 +4,4 @@ instance
   (FromJSON (c 'Partial), Monoid (c 'Partial))
   => FromJSON (PtPostProcessConfig c)
   where
+    parseJSON = parser