+++ source/test-suite/fixtures/InstanceHeadInput.hs
@@ -1,3 +1,4 @@
 {-# LANGUAGE DataKinds #-}
 instanceHeadAnchor = ()
 instance (FromJSON (c 'Partial), Monoid (c 'Partial)) => FromJSON (PtPostProcessConfig c) where
+  parseJSON = parser